+++ 03-HAL/TEMP/TEMP_program.c
@@ -1,6 +1,5 @@
 #include "../../01-LIB/STD_TYPES.h"
 
-//#include "../ADC_interface.h"
 #include "../../02-MCAL/ADC/ADC_interface.h"
 #include "TEMP_private.h"
 #include "TEMP_config.h"
@@ -16,7 +15,7 @@ uint8_t TEMP_u8GetReading(void)
 {
 	uint8_t local_u8TEMP;
 	uint16_t local_u16MiliVolt;
-	uint8_t local_u8ADC_Read;
+	uint16_t local_u8ADC_Read;
 	ADC_u8StartConversionSynch(ADC_Channel,&local_u8ADC_Read);
 	local_u16MiliVolt=(uint16_t)(((uint32_t)local_u8ADC_Read*Vref)/TWO_P_RESOLUTION);
 	local_u8TEMP=local_u16MiliVolt/10;
